feat(context-menu): Swarm copy menu, agent-pane paste menu, and app-wide fix for the same gap
The pane-body context menu (block/pane-actions.ts) only ever offers Copy-on-
selection and Paste for `view: "term"` panes, so every <input>/<textarea>
living directly in a non-terminal pane body (not inside a portalled Modal,
which already worked) had no way to paste via right-click - it got hijacked
into a useless disabled-Copy menu instead. Swarm had no context menu at all,
and its rows are user-select:none, so right-click did nothing.

Fix: extracted app.tsx's root Cut/Copy/Paste fallback into a shared,
exported showTextInputContextMenu() in store/contextmenu.ts, and attached it
directly to every affected input/textarea found in a full sweep (agent
decision/question panels, browser address bar, skill/MCP/memory/brain
manager forms, drone node editor fields) - stopPropagation on the element
itself always wins over the pane-body handler. Swarm's primary agent row
gets a native ContextMenuModel Copy menu (name/block ID), following the
AgentPicker.tsx per-row pattern.

See docs/specs/REPORT_CONTEXT_MENU_GAP_AUDIT_2026_08_07.md for the full
gap inventory, including what's intentionally left for a follow-up pass
(list/table copy-id polish, low-value single-line inputs).
