+++ .changesets/1786158722-feat-context-menu-swarm-copy-menu-agent-pane-composer-paste--7b7y.md
@@ -1,0 +1,5 @@
+---
+type: minor
+---
+
+feat(context-menu): Swarm copy menu, agent-pane composer paste menu, and shared Cut/Copy/Paste fix for every in-pane text input missing it
