+++ frontend/app/store/contextmenu.ts
@@ -1,7 +1,9 @@
 // Copyright 2025-2026, AgentMux Corp.
 // SPDX-License-Identifier: Apache-2.0
 
-import { atoms, getApi } from "./global";
+import { atoms, getApi, openLink } from "./global";
+import { readText as clipboardReadText } from "@/util/clipboard";
+import * as util from "@/util/util";
 
 class ContextMenuModelType {
     handlers: Map<string, () => void> = new Map(); // id -> handler
@@ -61,4 +63,91 @@ class ContextMenuModelType {
 
 const ContextMenuModel = new ContextMenuModelType();
 
-export { ContextMenuModel };
+// ── Shared Cut/Copy/Paste text-input menu ────────────────────────────────
+//
+// Any pane body (block/blockframe.tsx's onBodyContextMenu) intercepts
+// right-click before it can reach the app-root fallback in app.tsx, and
+// replaces it with a generic Copy-on-selection-only menu that never offers
+// Paste (except for `view: "term"` panes). That leaves every <input>/
+// <textarea> living directly in a pane body (not inside a portalled Modal,
+// which escapes the pane body entirely) with no way to paste via right-click.
+// This is the fix: attach `onContextMenu={showTextInputContextMenu}` directly
+// on the element. `stopPropagation` keeps the event from ever reaching
+// blockframe's handler, so this always wins for the element it's attached
+// to, matching what already works correctly for modal inputs.
+function isContentEditableBeingEdited(): boolean {
+    const activeElement = document.activeElement;
+    return (
+        activeElement != null &&
+        activeElement.getAttribute("contenteditable") !== null &&
+        activeElement.getAttribute("contenteditable") !== "false"
+    );
+}
+
+function canEnablePaste(): boolean {
+    const activeElement = document.activeElement;
+    return activeElement?.tagName === "INPUT" || activeElement?.tagName === "TEXTAREA" || isContentEditableBeingEdited();
+}
+
+function canEnableCopy(): boolean {
+    const sel = window.getSelection();
+    return !util.isBlank(sel?.toString());
+}
+
+function canEnableCut(): boolean {
+    const sel = window.getSelection();
+    if (document.activeElement?.classList.contains("xterm-helper-textarea")) {
+        return false;
+    }
+    return !util.isBlank(sel?.toString()) && canEnablePaste();
+}
+
+async function getClipboardURL(): Promise<URL | null> {
+    try {
+        const clipboardText = await clipboardReadText();
+        if (clipboardText == null) {
+            return null;
+        }
+        const url = new URL(clipboardText);
+        if (!url.protocol.startsWith("http")) {
+            return null;
+        }
+        return url;
+    } catch (e) {
+        return null;
+    }
+}
+
+async function showTextInputContextMenu(e: MouseEvent): Promise<void> {
+    e.preventDefault();
+    e.stopPropagation();
+    const canPaste = canEnablePaste();
+    const canCopy = canEnableCopy();
+    const canCut = canEnableCut();
+    const clipboardURL = await getClipboardURL();
+    if (!canPaste && !canCopy && !canCut && !clipboardURL) {
+        return;
+    }
+    let menu: ContextMenuItem[] = [];
+    if (canCut) {
+        menu.push({ label: "Cut", role: "cut" });
+    }
+    if (canCopy) {
+        menu.push({ label: "Copy", role: "copy" });
+    }
+    if (canPaste) {
+        menu.push({ label: "Paste", role: "paste" });
+    }
+    if (clipboardURL) {
+        menu.push({ type: "separator" });
+        menu.push({
+            label: "Open Clipboard URL (" + clipboardURL.hostname + ")",
+            click: () => {
+                openLink(clipboardURL.toString());
+            },
+        });
+    }
+    ContextMenuModel.showContextMenu(menu, e);
+}
+
+export { ContextMenuModel, showTextInputContextMenu };
